change type to Integer

+++ src/main/java/com/example/Tickets/TicketRepository.java
@@ -2,6 +2,6 @@ package com.example.Tickets;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepository extends JpaRepository<Ticket, Long> {
+public interface TicketRepository extends JpaRepository<Ticket, Integer> {
     Ticket findByMessage(String message);
 }
